refactor(grading): fix inverted repository -> services dependency
repositories/view_base.py imported aggregate_grading_status UP from
services.course_member_grading_stats -- a bottom-layer repository depending on a
service. The function is pure (no DB/service deps), so move it down to the util
layer (utils/grading_status.py) where both layers can depend on it.

services.course_member_grading_stats re-exports it, so the other callers
(analytics.grading_repository, business_logic.course_member_gradings, and the
module's own process_hierarchical_stats) are unchanged.

Note: two further repository->services inversions remain and need their own
design decision -- they import services that have real DB/storage dependencies,
not pure helpers, so they can't just be moved down:
  - repositories/course_member_gradings_view.py -> services.course_member_grading_read
  - repositories/view_mappers.py -> services.result_storage
