+++ computor-backend/src/computor_backend/services/course_member_grading_stats.py
@@ -1,23 +1,12 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from typing import Any, Iterable
+from typing import Any
 
-
-def aggregate_grading_status(
-    statuses: Iterable[str | None],
-    default: str | None = "not_reviewed",
-) -> str | None:
-    valid_statuses = [status for status in statuses if status is not None]
-    if not valid_statuses:
-        return default
-    if "correction_necessary" in valid_statuses:
-        return "correction_necessary"
-    if "improvement_possible" in valid_statuses:
-        return "improvement_possible"
-    if all(status == "corrected" for status in valid_statuses):
-        return "corrected"
-    return "not_reviewed"
+# Re-exported from the util layer so repositories can use it without an upward
+# repository -> services import (the canonical definition lives in
+# computor_backend.utils.grading_status).
+from computor_backend.utils.grading_status import aggregate_grading_status
 
 
 def process_hierarchical_stats(
